Refactor OperationsTimeline, chart hover state
